Scenes

+++ _OctaviCopter/Assets/Scripts/Database/UserProgress.cs
@@ -6,8 +6,8 @@ using SQLite4Unity3d;
 [Table ("UserProgress")]
 public class UserProgress 
 {
-   [PrimaryKey] public long LevelID { get; set; }
-    public long UserID { get; set; }    // FK on User
+   [PrimaryKey] public int LevelID { get; set; }
+    public int UserID { get; set; }    // FK on User
     public int LevelNumber { get; set; }
     public string LevelName { get; set; }
 }
